requirements.txt
certifi==2020.6.20
cffi==1.14.0
cryptography==3.3.2
decorator==4.4.2
future==0.18.2
pycparser==2.20
python-telegram-bot==12.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4
six==1.15.0
tornado==6.0.4
python-i18n==0.3.9
SQLAlchemy~=1.3.24